[[Foo a := b]] semantics supported

+++ src/builtin/das_qsort_r.h
@@ -194,7 +194,7 @@ inline void das_qsort_r(void *base, uint32_t nel, uint32_t width, Compare cmp)
             shr(p, 2);
             pshift += 2;
         } else {
-            if(lp[pshift - 1] >= high - head) {
+            if(lp[pshift - 1] >= uint32_t(high - head)) {
                 trinkle(head, width, cmp, p, pshift, 0, lp);
             } else {
                 sift(head, width, cmp, pshift, lp);
